refactor(comparison): consolidate color comparison internals
What changed:
- split comparison presentation and SwiftUI views from the calculation API
- centralize pure HSL and WCAG sRGB calculations for comparison and accessibility callers
- clarify CIEDE2000 and legacy fallback names while removing narration comments
- centralize perceptual metric display labels

Why:
- keep comparison responsibilities focused and prevent duplicated formulas or misleading terminology from drifting

Alternatives considered:
- leaving the duplicated conversion code in place was rejected because it made equivalent calculations evolve independently
